Move enhancements into base classes
`publisher` used to reopen and extend classes defined by
`govuk_content_models`. Given that these are now incorporated into
`publisher` itself, the classes themselves can be edited instead of
reopening them.
