--- lib/enhancements/artefact.rb
@@ -1,57 +1,0 @@
-require "artefact"
-
-class Artefact
-  before_destroy :discard_publishing_api_draft
-
-  MULTIPART_FORMATS = %w(guide local_transaction licence simple_smart_answer)
-
-  def self.multipart_formats
-    where(kind: { '$in' => MULTIPART_FORMATS })
-  end
-
-  def self.archived
-    where(state: 'archived')
-  end
-
-  def self.with_redirect
-    where(:redirect_url.nin => [nil, ""])
-  end
-
-  def latest_edition
-    Edition
-      .where(panopticon_id: id)
-      .order(version_number: :desc)
-      .first
-  end
-
-  def latest_edition_id
-    edition = latest_edition
-    edition.id.to_s if edition
-  end
-
-  def update_from_edition(edition)
-    update_attributes(
-      state: state_from_edition(edition),
-      description: edition.overview,
-      public_timestamp: edition.public_updated_at
-    )
-  end
-
-  def downtime
-    Downtime.for(self)
-  end
-
-private
-
-  def state_from_edition(edition)
-    case edition.state
-    when 'published' then 'live'
-    when 'archived' then 'archived'
-    else 'draft'
-    end
-  end
-
-  def discard_publishing_api_draft
-    Services.publishing_api.discard_draft(self.content_id)
-  end
-end
